+++ lib/substring_highlight.dart
@@ -2,11 +2,15 @@ library substring_highlight;
 
 import 'package:flutter/material.dart';
 
+const int int64MaxValue = 9223372036854775807; //HACK
+
 /// Widget that renders a string with sub-string highlighting.
 class SubstringHighlight extends StatelessWidget {
   SubstringHighlight({
     required this.text,
-    required this.term,
+    this.term,
+    this.terms,
+    this.caseSensitive = false,
     this.maxLines,
     this.overflow = TextOverflow.clip,
     this.textStyle = const TextStyle(
@@ -15,18 +19,24 @@ class SubstringHighlight extends StatelessWidget {
     this.textStyleHighlight = const TextStyle(
       color: Colors.red,
     ),
-  });
+  }) : assert(term != null || terms != null);
 
-  /// The String searched by {SubstringHighlight.term}.
+  /// The String searched by {SubstringHighlight.term} and/or {SubstringHighlight.terms} array.
   final String text;
 
-  /// The sub-string that is highlighted inside {SubstringHighlight.text}.
-  final String term;
+  /// The sub-string that is highlighted inside {SubstringHighlight.text}.  (Either term or terms must be passed.  If both are passed they are combined.)
+  final String? term;
+
+  /// The array of sub-strings that are highlighted inside {SubstringHighlight.text}.  (Either term or terms must be passed.  If both are passed they are combined.)
+  final List<String>? terms;
+
+  /// By default the search terms are case insensitive.  Pass false to force case sensitive matches.
+  final bool caseSensitive;
 
   /// The {TextStyle} of the {SubstringHighlight.text} that isn't highlighted.
   final TextStyle textStyle;
 
-  /// The {TextStyle} of the {SubstringHighlight.term}s found.
+  /// The {TextStyle} of the {SubstringHighlight.term}/{SubstringHighlight.ters} matched.
   final TextStyle textStyleHighlight;
 
   /// How visual overflow should be handled.
@@ -42,34 +52,71 @@ class SubstringHighlight extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    if (term.isEmpty) {
-      return Text(text,
-          maxLines: maxLines, overflow: overflow, style: textStyle);
-    } else {
-      String termLC = term.toLowerCase();
+    final String textLC = caseSensitive ? text : text.toLowerCase();
 
-      List<InlineSpan> children = [];
-      List<String> spanList = text.toLowerCase().split(termLC);
-      int i = 0;
-      for (var v in spanList) {
-        if (v.isNotEmpty) {
-          children.add(TextSpan(
-              text: text.substring(i, i + v.length), style: textStyle));
-          i += v.length;
-        }
-        if (i < text.length) {
-          children.add(TextSpan(
-              text: text.substring(i, i + term.length),
-              style: textStyleHighlight));
-          i += term.length;
+    // if both term and terms array are passed then combine
+    final List<String> termList = [term ?? '', ...(terms ?? [])];
+
+    // must remove empty search terms because they cause infinite loops
+    final List<String> termListLC = termList
+        .where((s) => s.isNotEmpty)
+        .map((s) => caseSensitive ? s : s.toLowerCase())
+        .toList();
+
+    List<InlineSpan> children = [];
+
+    int start = 0;
+    int idx = 0;
+    while (idx < textLC.length) {
+      addNonHighlight(int end) => children
+          .add(TextSpan(text: text.substring(start, end), style: textStyle));
+
+      // find index of term that's closest to current idx position
+      int iNearest = -1;
+      int idxNearest = int64MaxValue;
+      for (int i = 0; i < termListLC.length; i++) {
+        // print('idx=$idx i=$i $termsLC');
+        int at;
+        if ((at = textLC.indexOf(termListLC[i], idx)) >= 0) {
+          // print('term #$i found at=$at (${termListLC[i]})');
+          if (at < idxNearest) {
+            iNearest = i;
+            idxNearest = at;
+          }
         }
       }
 
-      return RichText(
-          maxLines: maxLines,
-          overflow: overflow,
-          text: TextSpan(children: children, style: textStyle),
-          textScaleFactor: MediaQuery.of(context).textScaleFactor);
+      if (iNearest >= 0) {
+        // found one of the terms at or after idx
+        // iNearest is the index of the closest term at or after idx that matches
+
+        // print('iNearest=$iNearest @ $idxNearest');
+        if (start < idxNearest) {
+          // we found a match BUT FIRST output non-highlighted text that comes BEFORE this match
+          addNonHighlight(idxNearest);
+          start = idxNearest;
+        }
+
+        // output the match using desired highlighting
+        int termLen = termListLC[iNearest].length;
+        children.add(TextSpan(
+            text: text.substring(start, idxNearest + termLen),
+            style: textStyleHighlight));
+        start = idx = idxNearest + termLen;
+      } else {
+        // if none match at all (ever!)
+        // --or--
+        // one or more matches but during this iteration there are NO MORE matches
+        // in either case, add reminder of text as non-highlighted text
+        addNonHighlight(textLC.length);
+        break;
+      }
     }
+
+    return RichText(
+        maxLines: maxLines,
+        overflow: overflow,
+        text: TextSpan(children: children, style: textStyle),
+        textScaleFactor: MediaQuery.of(context).textScaleFactor);
   }
 }